+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+ - `--horizontal/-h` and `--vertical/-v` flags for `bot send video` has been
+ changed to `--width/-w` ad `--height/-h`.
+
 ## [v0.3.1] - 2019-09-22
 ### Changed
  - The description of `--horizontal` and `--vertical` options on `bot send video`